main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv
from routers import resume, video, audio
import os

load_dotenv()

app = FastAPI(title="Interview AI Backend")

cors_origins = os.getenv("CORS_ORIGINS", "http://localhost:3000")
allowed_origins = [origin.strip() for origin in cors_origins.split(",") if origin.strip()]

app.add_middleware(
    CORSMiddleware,
    allow_origins=allowed_origins or ["http://localhost:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)
app.include_router(resume.router)
app.include_router(video.router)
app.include_router(audio.router)
@app.get("/")
def health_check():
    return {"status": "Interview AI Backend Running"}

if __name__ == "__main__":
    # Hugging Face provides the PORT environment variable, usually 7860
    port = int(os.environ.get("PORT", 7860)) 
    uvicorn.run("main:app", host="0.0.0.0", port=port)