requirements.txt
groq==0.5.0
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
pdfplumber
easyocr
openai
openai-whisper
pydub
opencv-python
mediapipe
numpy
Pillow